:sunflower: Part way through refactor
